Make Resume a member function of ContinuatoionHandle; begin TransferControl
Make TransferControl implemented in terms of a setjmp-like opcode that
returns the continuation pointer if called normally, or zero if we're
resuming.
that way the IP in the interpreter will be correct, and we will have
locals that get populated with the answer or the continuation.
And we don't have to implement delegate invocation in the opcode.

On the other hand we will be running some managed code in the new
frame that we set up, so we will have to copy over some locals (ie we
need to duplicate the topmost InterpFrame into the new stack) so that
we can do the delegate call.