Fix logic bug in demo scheduler
When enqueuing a resumed computation when there's only one green
thread, the queue is empty, but the scheduler is already running
because we just yielded from it - don't kick off a new scheduler loop.

As part of debugging this, I moved the loop logic from an async Task
into a JS setTimeout loop.  TODO move it back